Assert that Vulkan array-getters return the same length
Originally introduced in [#489] this inserts the array-length equality
check everywhere else: in the supposedly invalid and inexistant event
where Vulkan suddenly returns less items (`count` has been modified)
than were originally queried through respective `_len()` functions (or
more likely: a slice of invalid length passed by the user) and some
elements at the end of the slice are left uninitialized, panic.

Wherever there is valid concern or possibility for this to happen - or
to circumvent assertions and panics altogether - mutable references to
mutable slices should be passed allowing the length to be promptly
updated.

[#489]: https://github.com/MaikKlein/ash/pull/489#discussion_r753020089

+++ ash/src/extensions/khr/get_surface_capabilities2.rs
@@ -66,7 +66,7 @@ impl GetSurfaceCapabilities2 {
             &mut count,
             out.as_mut_ptr(),
         );
-        assert_eq!(count, out.len() as u32);
+        assert_eq!(count as usize, out.len());
         err_code.result()
     }
