+++ ash/src/extensions/khr/maintenance4.rs
@@ -69,7 +69,7 @@ impl Maintenance4 {
             &mut count,
             out.as_mut_ptr(),
         );
-        assert_eq!(count, out.len() as u32);
+        assert_eq!(count as usize, out.len());
     }
 
     pub fn name() -> &'static CStr {
